@@ -1,5 +1,5 @@
 module github.com/go-cmd/cmd
 
-go 1.14
+go 1.17
 
-require github.com/go-test/deep v1.0.6
+require github.com/go-test/deep v1.0.7
